chore: move BitStorage classes into a separate folder
