+++ src/bitstorage/bin_bit_storage.h
