+++ src/bitstorage/bit_storage.h
