+++ src/bitstorage/hex_bit_storage.h
